// lavastore-pro/apps/api/src/modules/orders/order.schema.ts
import { z } from 'zod';

export const createOrderSchema = z.object({
  items: z.array(z.object({
    productId: z.string().uuid(),
    quantity: z.number().int().positive(),
  })).min(1, 'Pedido deve conter ao menos 1 item'),
  shippingAddress: z.object({
    street: z.string(),
    number: z.string(),
    complement: z.string().optional(),
    neighborhood: z.string(),
    city: z.string(),
    state: z.string(),
    zipCode: z.string(),
  }),
});

export type CreateOrderInput = z.infer<typeof createOrderSchema>;